Learn a lot about json today
Added in all the maps I got off the wiki to the locations files. I couldnt figure out how you were pulling from the lava-dome dungeon file to access the maps in a new file so I based all the images out of that one file till either I figure out how to make it work  or I'm given a lesson.

Starting to add in the logic for the items in the Lava Dome to get use to adding all the icons to each map.